fix signature

+++ steel-worldgen/src/density/traits.rs
@@ -84,15 +84,15 @@ pub trait DimensionNoises: Sized + Send + Sync {
     // ── Router functions ────────────────────────────────────────────────────
 
     /// Final density for terrain generation (positive = solid, negative = air).
-    fn router_final_density(&self, cache: &mut Self::ColumnCache, x: i32, y: i32, z: i32) -> f64;
+    fn router_final_density(&self, cache: &mut Self::ColumnCache, x: i32, y: i32, z: i32) -> f32;
 
     /// Depth from surface (used for terrain shaping).
-    fn router_depth(&self, cache: &mut Self::ColumnCache, x: i32, y: i32, z: i32) -> f64;
+    fn router_depth(&self, cache: &mut Self::ColumnCache, x: i32, y: i32, z: i32) -> f32;
 
     // ── Aquifer router functions ────────────────────────────────────────────
 
     /// Barrier noise for aquifer boundaries.
-    fn router_barrier(&self, cache: &mut Self::ColumnCache, x: i32, y: i32, z: i32) -> f64;
+    fn router_barrier(&self, cache: &mut Self::ColumnCache, x: i32, y: i32, z: i32) -> f32;
 
     /// Fluid level floodedness for aquifers.
     fn router_fluid_level_floodedness(
@@ -101,7 +101,7 @@ pub trait DimensionNoises: Sized + Send + Sync {
         x: i32,
         y: i32,
         z: i32,
-    ) -> f64;
+    ) -> f32;
 
     /// Fluid level spread for aquifers.
     fn router_fluid_level_spread(
@@ -110,38 +110,38 @@ pub trait DimensionNoises: Sized + Send + Sync {
         x: i32,
         y: i32,
         z: i32,
-    ) -> f64;
+    ) -> f32;
 
     /// Lava placement noise for aquifers.
-    fn router_lava(&self, cache: &mut Self::ColumnCache, x: i32, y: i32, z: i32) -> f64;
+    fn router_lava(&self, cache: &mut Self::ColumnCache, x: i32, y: i32, z: i32) -> f32;
 
     // ── Ore vein router functions ───────────────────────────────────────────
 
     /// Vein toggle (sign determines copper vs iron).
-    fn router_vein_toggle(&self, cache: &mut Self::ColumnCache, x: i32, y: i32, z: i32) -> f64;
+    fn router_vein_toggle(&self, cache: &mut Self::ColumnCache, x: i32, y: i32, z: i32) -> f32;
 
     /// Vein ridged noise for ore placement.
-    fn router_vein_ridged(&self, cache: &mut Self::ColumnCache, x: i32, y: i32, z: i32) -> f64;
+    fn router_vein_ridged(&self, cache: &mut Self::ColumnCache, x: i32, y: i32, z: i32) -> f32;
 
     /// Vein gap noise for ore vs filler placement.
-    fn router_vein_gap(&self, cache: &mut Self::ColumnCache, x: i32, y: i32, z: i32) -> f64;
+    fn router_vein_gap(&self, cache: &mut Self::ColumnCache, x: i32, y: i32, z: i32) -> f32;
 
     // ── Climate/biome router functions (Y-independent, cached) ──────────────
 
     /// Erosion value (cached in column cache).
-    fn router_erosion(&self, cache: &mut Self::ColumnCache, x: i32, y: i32, z: i32) -> f64;
+    fn router_erosion(&self, cache: &mut Self::ColumnCache, x: i32, y: i32, z: i32) -> f32;
 
     /// Continentalness value (cached in column cache).
-    fn router_continentalness(&self, cache: &mut Self::ColumnCache, x: i32, y: i32, z: i32) -> f64;
+    fn router_continentalness(&self, cache: &mut Self::ColumnCache, x: i32, y: i32, z: i32) -> f32;
 
     /// Temperature value (cached in column cache).
-    fn router_temperature(&self, cache: &mut Self::ColumnCache, x: i32, y: i32, z: i32) -> f64;
+    fn router_temperature(&self, cache: &mut Self::ColumnCache, x: i32, y: i32, z: i32) -> f32;
 
     /// Vegetation/humidity value (cached in column cache).
-    fn router_vegetation(&self, cache: &mut Self::ColumnCache, x: i32, y: i32, z: i32) -> f64;
+    fn router_vegetation(&self, cache: &mut Self::ColumnCache, x: i32, y: i32, z: i32) -> f32;
 
     /// Ridges/weirdness value (cached in column cache).
-    fn router_ridges(&self, cache: &mut Self::ColumnCache, x: i32, y: i32, z: i32) -> f64;
+    fn router_ridges(&self, cache: &mut Self::ColumnCache, x: i32, y: i32, z: i32) -> f32;
 
     /// Preliminary surface level (cached in column cache).
     fn router_preliminary_surface_level(
@@ -150,7 +150,7 @@ pub trait DimensionNoises: Sized + Send + Sync {
         x: i32,
         y: i32,
         z: i32,
-    ) -> f64;
+    ) -> f32;
 
     // ── Interpolation functions ─────────────────────────────────────────────
